refactor: Move merge protocol to goblin module and separate interface from implementation
This commit reorganizes the merge protocol code:

1. Separated interface from implementation:
   - Moved inline implementations from merge_prover.hpp to merge_prover.cpp
   - Moved inline implementations from merge_verifier.hpp to merge_verifier.cpp
   - This improves compilation times and code organization

2. Relocated merge protocol files to goblin module:
   - Moved merge_prover.{hpp,cpp} from ultra_honk/ to goblin/
   - Moved merge_verifier.{hpp,cpp} from ultra_honk/ to goblin/
   - The merge protocol is part of Goblin, so this better reflects the architecture

3. Updated all references:
   - Updated includes in mega_honk.test.cpp, goblin.hpp, goblin.cpp
   - Updated includes in stdlib tests and recursive verifiers
   - Updated ultra_honk CMakeLists.txt to depend on goblin module

All tests compile and pass successfully.
