+++ barretenberg/cpp/src/barretenberg/stdlib/goblin_verifier/goblin_recursive_verifier.hpp
@@ -6,9 +6,9 @@
 
 #pragma once
 #include "barretenberg/goblin/goblin.hpp"
+#include "barretenberg/goblin/merge_verifier.hpp"
 #include "barretenberg/stdlib/eccvm_verifier/eccvm_recursive_verifier.hpp"
 #include "barretenberg/stdlib/translator_vm_verifier/translator_recursive_verifier.hpp"
-#include "barretenberg/ultra_honk/merge_verifier.hpp"
 
 namespace bb::stdlib::recursion::honk {
